First Day Tutorial
First day tutorial sequence is mostly complete. I changed the dialogue so that it always advances automatically after a short time and doesn't interfere with other buttons. Only problem is, a bug occurs if you sell something to the boss too early, but I can't exactly figure out how to enable the inventory box mid dialogue which would be ideal. Also not sure how to advance to the next day automatically after finishing.